Release 1.5.3
- [+] **Add** Active mode to HitSelect.
    - Active mode will cancel attack action and keep swing action. It looks like "Legit KeepSprint", and may be flagged by Polar.
- [※] **Fix** JumpReset sometimes do impossible jump.
- [+] **Improve** TimerRange.
    - **Add** fov option. default is 180° it means only TimerRange if you are face the target.
    - **Add** ignoreTeammates option.
    - Now TimerRange didn't need KillAura anymore.
- [※] **Improve** StrafeTest mode in Speed.
    - Renamed "Strafe (Deprecated)" to "StrafeTest".
- [+] **Add** Latency compensation option to Water bucket.
    - if the value > 0, module will predict your future move and extend your reach distance.
- [※] **Fix** Block-In
    - fix direction

+++ src/main/java/keystrokesmod/module/impl/render/BreakProgress.java
@@ -19,7 +19,7 @@ public class BreakProgress extends Module {
     private ButtonSetting manual;
     private ButtonSetting bedAura;
     private String[] modes = new String[]{"Percentage", "Second", "Decimal"};
-    private float progress;
+    private double progress;
     private BlockPos block;
     private String progressStr;
 
@@ -76,7 +76,7 @@ public class BreakProgress extends Module {
             return;
         }
         if (bedAura.isToggled() && ModuleManager.bedAura != null && ModuleManager.bedAura.isEnabled() && ModuleManager.bedAura.breakProgress != 0.0f && ModuleManager.bedAura.currentBlock != null && !(BlockUtils.getBlock(ModuleManager.bedAura.currentBlock) instanceof BlockBed)) {
-            this.progress = Math.min(1.0f, ModuleManager.bedAura.breakProgress);
+            this.progress = Math.min(1.0, ModuleManager.bedAura.breakProgress);
             this.block = ModuleManager.bedAura.currentBlock;
             if (this.block == null) {
                 return;
